backend/requirements.txt
fastapi
uvicorn[standard]
groq
openai
python-dotenv
pydantic
gTTS
python-multipart
